my-jemo-app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.2",
    "fastapi>=0.136.0",
    "tortoise-orm>=0.25.4",
    "uvicorn>=0.45.0",
]

[tool.aerich]
tortoise_orm = "tortoise_conf.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
